Add files via upload
This repository contains a logistic regression model that predicts the acceptance of loan insurance offers based on data from UniversalBank. The model uses customer data such as age, income, and credit score to predict the likelihood of a customer accepting a loan insurance offer. The model is trained on historical data from UniversalBank and can be used to make predictions on new customer data.