Add files via upload
The Netflix dataset available on Kaggle provides information on various movies and TV shows available on the platform. The dataset contains details such as title, director, cast, genre, rating, and release year for each entry. With the help of this dataset, it is possible to determine the number of shows and movies that were released in the year 2000.